Avance

+++ python/yt_downloader.py
@@ -70,7 +70,7 @@ def get_songs(raw_string):
 if __name__=="__main__":
     
     normalizer = ffmpeg_normalize.FFmpegNormalize(audio_codec="mp3",dynamic=True,sample_rate=48000,progress=True)
- 
+    
     if ".txt" in sys.argv[1]:
         with open (sys.argv[1],encoding="UTF-8")as file:
             raw_songs= file.readline()
@@ -78,22 +78,28 @@ if __name__=="__main__":
         songs= get_songs(raw_songs)
     else:
         songs=[sys.argv[1]]
-    
+    just_one = len(songs) ==1
     for index,song in enumerate(songs) :
         print("LOG:"+str(index+1)  )
         auxi = re.sub("[\\/:*?\"<>|]","#",song)
         download_video("ytsearch:"+auxi)
-       
-        normalizer.add_media_file("./python/canciones/"+auxi+".mp3","./python/normalizadas/"+auxi+".mp3")
+        if not just_one:
+            normalizer.add_media_file("./python/canciones/"+auxi+".mp3","./python/normalizadas/"+auxi+".mp3")
     #print(f"LOG:Normalizando las {len(songs)} canciones")
-    print("Normalizando canciones...")    
-    normalizer.run_normalization()
+    print("Normalizando canciones...")
+    if not just_one:    
+        normalizer.run_normalization()
     for file in os.listdir("./python/canciones/"):
-        os.remove("./python/canciones/"+file)
-    if ".txt" in sys.argv[1]:
-        nombre=os.path.basename(sys.argv[1]).split(".")[0]  
+        if just_one:
+            os.rename("./python/canciones/"+file,"./python/normalizadas/"+file)
+        else:       
+            os.remove("./python/canciones/"+file)
+    if just_one:
+        nombre =re.sub("[\\/:*?\"<>|]","#",songs[0])
+        
     else:
-        nombre =re.sub("[\\/:*?\"<>|]","#",sys.argv[1])
+        nombre=os.path.basename(sys.argv[1]).split(".")[0] 
+        
     zip = zipfile.ZipFile("./songs_list/"+nombre+".zip", "w", zipfile.ZIP_STORED,compresslevel=1)
     for file in os.listdir("./python/normalizadas/"):
         if ".git" in file:
